chore: clean up reused identity util
On-behalf-of: SAP <matthias.bruns@sap.com>
