+++ bindings/go/wget/access/access.go
@@ -6,6 +6,7 @@ import (
 
 	descruntime "ocm.software/open-component-model/bindings/go/descriptor/runtime"
 	"ocm.software/open-component-model/bindings/go/runtime"
+	"ocm.software/open-component-model/bindings/go/wget/internal/identity"
 	accessspec "ocm.software/open-component-model/bindings/go/wget/spec/access"
 	"ocm.software/open-component-model/bindings/go/wget/spec/access/v1"
 )
@@ -33,24 +34,5 @@ func (w *WgetAccess) GetResourceCredentialConsumerIdentity(ctx context.Context, 
 		return nil, fmt.Errorf("error converting resource access spec: %w", err)
 	}
 
-	return CredentialConsumerIdentity(wget.URL)
-}
-
-// CredentialConsumerIdentity resolves the credential consumer identity for the given
-// wget URL. It is shared by the wget access type and the wget input method so that
-// credentials configured for a host resolve identically whether the resource is
-// declared via an access spec or an input spec.
-func CredentialConsumerIdentity(url string) (runtime.Identity, error) {
-	if url == "" {
-		return nil, fmt.Errorf("url is required")
-	}
-
-	identity, err := runtime.ParseURLToIdentity(url)
-	if err != nil {
-		return nil, fmt.Errorf("error parsing wget URL to identity: %w", err)
-	}
-
-	identity.SetType(accessspec.V1VersionedType)
-
-	return identity, nil
+	return identity.CredentialConsumerIdentity(wget.URL)
 }
