+++ bindings/go/wget/doc.go
@@ -1,4 +1,5 @@
-// Package wget provides HTTP(S) download access for OCM resources.
+// Package wget provides HTTP(S) access to OCM resources, both as an access type
+// and as a component-constructor input method.
 //
 // It implements the "Wget" access type: a resource whose bytes are fetched
 // from an HTTP or HTTPS endpoint described by a
@@ -30,6 +31,24 @@
 // [ocm.software/open-component-model/bindings/go/wget/access.WgetAccess] derives
 // the credential consumer identity from a resource's URL, so a credential
 // resolver can look up matching credentials before the download.
+//
+// It also implements the "wget" constructor input method in
+// [ocm.software/open-component-model/bindings/go/wget/input.InputMethod], which
+// references an HTTP/S URL declared on a resource in a component-constructor.yaml
+// through a
+// [ocm.software/open-component-model/bindings/go/wget/spec/input/v1.Wget] input
+// spec. The input spec carries the same request details as the access spec and
+// supports two output modes:
+//
+//   - local blob (default): the content is downloaded during construction and
+//     stored as a local blob, making the component version self-contained.
+//   - access spec (the input sets Reference): the content is not downloaded; the
+//     resource is stored with a wget access spec pointing at the URL, so it is
+//     fetched lazily when the resource is later accessed.
+//
+// The input method shares the download transport, credential handling and size
+// limiting used by the access type, so both behave identically for a given URL
+// and credentials, and both derive the credential consumer identity from the URL.
 //
 // The wire types are each registered in their package scheme for typed
 // conversion. Both the versioned (wget/v1) and unversioned (wget) type names are
