+++ bindings/go/wget/input/method.go
@@ -10,8 +10,8 @@ import (
 	httpclient "ocm.software/open-component-model/bindings/go/http"
 	httpv1alpha1 "ocm.software/open-component-model/bindings/go/http/spec/config/v1alpha1"
 	"ocm.software/open-component-model/bindings/go/runtime"
-	"ocm.software/open-component-model/bindings/go/wget/access"
 	"ocm.software/open-component-model/bindings/go/wget/internal/download"
+	"ocm.software/open-component-model/bindings/go/wget/internal/identity"
 	"ocm.software/open-component-model/bindings/go/wget/repository"
 	accessspec "ocm.software/open-component-model/bindings/go/wget/spec/access"
 	accessv1 "ocm.software/open-component-model/bindings/go/wget/spec/access/v1"
@@ -21,14 +21,9 @@ import (
 
 var _ constructor.ResourceInputMethod = (*InputMethod)(nil)
 
-// InputMethod implements the ResourceInputMethod interface for wget-based inputs.
+// InputMethod implements the [constructor.ResourceInputMethod] interface for wget-based inputs.
 // It downloads a resource from an HTTP/S URL declared in the component constructor
 // and returns it as a local blob to be stored in the component version.
-//
-// The download transport, credential handling and size limiting are shared with the
-// wget access type via the internal download package: this method converts the wget
-// input specification into a download request, while the access type converts the
-// wget access specification into the same request.
 type InputMethod struct {
 	// HTTPConfig configures the HTTP client (timeouts, retries, TLS, routing) used for
 	// downloads. When nil, a default client is used.
@@ -51,14 +46,13 @@ func (i *InputMethod) GetResourceCredentialConsumerIdentity(_ context.Context, r
 		return nil, fmt.Errorf("error converting resource input spec: %w", err)
 	}
 
-	return access.CredentialConsumerIdentity(wget.URL)
+	return identity.CredentialConsumerIdentity(wget.URL)
 }
 
 // ProcessResource turns a wget input specification into a resource input method result.
 //
-// In the default local-blob mode it downloads the resource and returns it as local blob data;
-// the download is delegated to the shared wget download logic, so behavior matches the wget
-// access type exactly. When the input sets AsAccess, it instead returns a processed resource
+// In the default local-blob mode it downloads the resource and returns it as local blob data.
+// When the input sets Reference, it instead returns a processed resource
 // carrying a wget access specification pointing at the URL, without downloading anything.
 func (i *InputMethod) ProcessResource(ctx context.Context, resource *constructorruntime.Resource, credentials runtime.Typed) (*constructor.ResourceInputMethodResult, error) {
 	wget := v1.Wget{}
@@ -72,7 +66,7 @@ func (i *InputMethod) ProcessResource(ctx context.Context, resource *constructor
 
 	// Access-spec mode: do not download. Store the resource with a wget access specification
 	// pointing at the URL so the content is resolved lazily when the resource is accessed.
-	if wget.AsAccess {
+	if wget.Reference {
 		remoteResource, err := i.createRemoteResourceAccess(resource, wget)
 		if err != nil {
 			return nil, fmt.Errorf("error creating remote resource access: %w", err)
@@ -124,18 +118,11 @@ func (i *InputMethod) createRemoteResourceAccess(resource *constructorruntime.Re
 		Body:       wget.Body,
 		NoRedirect: wget.NoRedirect,
 	}
+
 	if _, err := accessspec.Scheme.DefaultType(wgetAccess); err != nil {
 		return nil, fmt.Errorf("error setting default type for wget access: %w", err)
 	}
-
-	// Encode the access into a runtime.Raw so downstream descriptor encoding does not depend on
-	// the wget access type being registered in the target repository's scheme. This mirrors how
-	// accesses declared directly in the component constructor are represented.
-	raw := &runtime.Raw{}
-	if err := accessspec.Scheme.Convert(wgetAccess, raw); err != nil {
-		return nil, fmt.Errorf("error encoding wget access: %w", err)
-	}
-	resource.Access = raw
+	resource.Access = wgetAccess
 
 	return resource, nil
 }
